Attempt at fixing https://github.com/firejump/vscode-frame-rainbow/issues/1
What is currently called "outer frame decorator" should actually be applied
to the inner frame non-whitespace decoration as well. Otherwise both
the text AND the newline character have additional after.contentText rendered.
These decorations are effectively overlapping and thus click events can reach
either one, resulting in the cursor being incorrectly put before the last character
of the line in some cases.

This is a WIP of the fix, just using the correct decorator; if this proves
to works well then we should rename inner and outer frame decorators, maybe
"textDecorator" and "eolDecorator"?